services/data-collector/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database Drivers
clickhouse-driver==0.2.6
psycopg2-binary==2.9.9
redis==4.6.0

# HTTP Client
httpx==0.25.1
aiohttp==3.9.1
requests==2.31.0

# Data Validation
pydantic==2.5.0

# Background Jobs & Scheduling
celery==5.3.4
celery[redis]==5.3.4
flower==2.0.1

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3
python-json-logger==2.0.7

# LLM & AI
groq==0.4.1

# Cloud SDKs
boto3==1.34.0  # AWS
google-cloud-billing==1.12.0  # GCP
google-cloud-monitoring==2.18.0  # GCP
google-cloud-compute==1.15.0  # GCP
azure-mgmt-costmanagement==4.0.1  # Azure
azure-mgmt-monitor==6.0.2  # Azure
azure-mgmt-compute==30.5.0  # Azure
azure-identity==1.15.0  # Azure
